Feature test IIIF viewer output
